rename few class and add propertis config
