+++ adapter/runtimer/src/main/java/org/apache/rocketmq/eventbridge/adapter/runtimer/boot/EventBusListener.java
@@ -20,7 +20,7 @@ package org.apache.rocketmq.eventbridge.adapter.runtimer.boot;
 import io.openmessaging.connector.api.data.ConnectRecord;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.rocketmq.eventbridge.adapter.runtimer.boot.listener.EventSubscriber;
-import org.apache.rocketmq.eventbridge.adapter.runtimer.boot.listener.ListenerFactory;
+import org.apache.rocketmq.eventbridge.adapter.runtimer.boot.listener.CirculatorContext;
 import org.apache.rocketmq.eventbridge.adapter.runtimer.common.ServiceThread;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -36,12 +36,12 @@ public class EventBusListener extends ServiceThread {
 
     private static final Logger logger = LoggerFactory.getLogger(EventBusListener.class);
 
-    private ListenerFactory listenerFactory;
+    private CirculatorContext circulatorContext;
 
     private EventSubscriber eventSubscriber;
 
-    public EventBusListener(ListenerFactory listenerFactory, EventSubscriber eventSubscriber) {
-        this.listenerFactory = listenerFactory;
+    public EventBusListener(CirculatorContext circulatorContext, EventSubscriber eventSubscriber) {
+        this.circulatorContext = circulatorContext;
         this.eventSubscriber = eventSubscriber;
     }
 
@@ -54,7 +54,7 @@ public class EventBusListener extends ServiceThread {
                     this.waitForRunning(1000);
                     continue;
                 }
-                listenerFactory.offerEventRecords(recordList);
+                circulatorContext.offerEventRecords(recordList);
             }catch (Exception exception) {
                 logger.error(getServiceName() + " - event bus pull record exception, stackTrace - ", exception);
             }
