+++ adapter/runtimer/src/main/java/org/apache/rocketmq/eventbridge/adapter/runtimer/boot/listener/RocketMQEventSubscriber.java
@@ -14,9 +14,11 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
+
 package org.apache.rocketmq.eventbridge.adapter.runtimer.boot.listener;
 
 import com.alibaba.fastjson.JSON;
+import com.google.common.base.Splitter;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import io.openmessaging.KeyValue;
@@ -29,19 +31,23 @@ import org.apache.commons.collections.MapUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.rocketmq.client.consumer.DefaultLitePullConsumer;
 import org.apache.rocketmq.client.exception.MQClientException;
+import org.apache.rocketmq.common.UtilAll;
 import org.apache.rocketmq.common.message.MessageExt;
+import org.apache.rocketmq.common.message.MessageQueue;
 import org.apache.rocketmq.eventbridge.adapter.runtimer.common.entity.TargetRunnerConfig;
 import org.apache.rocketmq.eventbridge.adapter.runtimer.common.enums.RefreshTypeEnum;
 import org.apache.rocketmq.eventbridge.adapter.runtimer.config.RuntimerConfigDefine;
+import org.apache.rocketmq.eventbridge.adapter.runtimer.service.TargetRunnerConfigObserver;
+import org.apache.rocketmq.eventbridge.exception.EventBridgeException;
+import org.apache.rocketmq.remoting.common.RemotingUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.core.io.support.PropertiesLoaderUtils;
 import org.springframework.util.CollectionUtils;
 
 import java.nio.charset.StandardCharsets;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
+import java.util.*;
+import java.util.stream.Collectors;
 
 /**
  * RocketMQ implement event subscriber
@@ -50,24 +56,33 @@ public class RocketMQEventSubscriber extends EventSubscriber {
 
     private static final Logger logger = LoggerFactory.getLogger(RocketMQEventSubscriber.class);
 
-    private ListenerFactory listenerFactory;
-
     private DefaultLitePullConsumer pullConsumer;
 
-    private Integer DEFAULT_PULL_TIME_OUT = 3000;
+    private TargetRunnerConfigObserver runnerConfigObserver;
 
-    public RocketMQEventSubscriber(ListenerFactory listenerFactory) {
-        this.listenerFactory = listenerFactory;
+    private Integer pullTimeOut;
+
+    private String namesrvAddr;
+
+    private static final String SEMICOLON = ";";
+
+    private static final String SYS_DEFAULT_GROUP = "event-bridge-default-group";
+
+    public static final String QUEUE_OFFSET = "queueOffset";
+
+    public RocketMQEventSubscriber(TargetRunnerConfigObserver runnerConfigObserver) {
+        this.runnerConfigObserver = runnerConfigObserver;
+        this.initMqProperties();
         this.initPullConsumer();
     }
 
     @Override
     public void refresh(TargetRunnerConfig targetRunnerConfig, RefreshTypeEnum refreshTypeEnum) {
         if(Objects.isNull(pullConsumer)){
-            pullConsumer = listenerFactory.initDefaultMQPullConsumer();
+            pullConsumer = initDefaultMQPullConsumer();
             return;
         }
-        Set<String> currentTopics = listenerFactory.parseTopicsByRunnerConfigs(Sets.newHashSet(targetRunnerConfig));
+        Set<String> currentTopics = parseTopicsByRunnerConfigs(Sets.newHashSet(targetRunnerConfig));
         for (String topic : currentTopics){
             switch (refreshTypeEnum){
                 case ADD:
@@ -85,7 +100,7 @@ public class RocketMQEventSubscriber extends EventSubscriber {
 
     @Override
     public List<ConnectRecord> pull() {
-        List<MessageExt> messageExts = pullConsumer.poll(DEFAULT_PULL_TIME_OUT);
+        List<MessageExt> messageExts = pullConsumer.poll(pullTimeOut);
         if (CollectionUtils.isEmpty(messageExts)) {
             logger.info("consumer poll message empty , consumer - {}", JSON.toJSONString(pullConsumer));
             return null;
@@ -107,10 +122,102 @@ public class RocketMQEventSubscriber extends EventSubscriber {
     }
 
     /**
+     * parse topics by specific target runner configs
+     * @param targetRunnerConfigs
+     * @return
+     */
+    public Set<String> parseTopicsByRunnerConfigs(Set<TargetRunnerConfig> targetRunnerConfigs){
+        if(org.apache.commons.collections.CollectionUtils.isEmpty(targetRunnerConfigs)){
+            logger.warn("target runner config is empty, parse to topic failed!");
+            return null;
+        }
+        Set<String> listenTopics = Sets.newHashSet();
+        for(TargetRunnerConfig runnerConfig : targetRunnerConfigs){
+            List<Map<String,String>> runnerConfigMap = runnerConfig.getComponents();
+            if(org.apache.commons.collections.CollectionUtils.isEmpty(runnerConfigMap)){
+                continue;
+            }
+            listenTopics.addAll(runnerConfigMap.stream().map(item->item.get(RuntimerConfigDefine.CONNECT_TOPICNAME)).collect(Collectors.toSet()));
+        }
+        return listenTopics;
+    }
+
+    /**
+     * init rocketmq ref config
+     */
+    private void initMqProperties() {
+        try {
+            Properties properties = PropertiesLoaderUtils.loadAllProperties("runtimer.properties");
+            namesrvAddr = properties.getProperty("rocketmq.namesrvAddr");
+            pullTimeOut = Integer.valueOf(properties.getProperty("rocketmq.consumer.pullTimeOut"));
+        }catch (Exception exception){
+
+        }
+
+    }
+
+    /**
      * init rocket mq pull consumer
      */
     private void initPullConsumer() {
-        pullConsumer = listenerFactory.initDefaultMQPullConsumer();
+        pullConsumer = initDefaultMQPullConsumer();
+    }
+
+    /**
+     * first init default rocketmq pull consumer
+     * @return
+     */
+    public DefaultLitePullConsumer initDefaultMQPullConsumer () {
+        Set<TargetRunnerConfig> targetRunnerConfigs = runnerConfigObserver.getTargetRunnerConfig();
+        Set<String> topics = parseTopicsByRunnerConfigs(targetRunnerConfigs);
+        DefaultLitePullConsumer consumer = new DefaultLitePullConsumer();
+        consumer.setConsumerGroup(createGroupName(SYS_DEFAULT_GROUP));
+        consumer.setNamesrvAddr(namesrvAddr);
+        try {
+            for(String topic : topics){
+                consumer.subscribe(topic, "*");
+            }
+            consumer.start();
+        } catch (Exception exception) {
+            logger.error("init default pull consumer exception, topic -" + topics.toString() + "-stackTrace-", exception);
+            throw new EventBridgeException(" init rocketmq consumer failed");
+        }
+        return consumer;
+    }
+
+    private String createGroupName(String prefix) {
+        StringBuilder sb = new StringBuilder();
+        sb.append(prefix).append("-");
+        sb.append(RemotingUtil.getLocalAddress()).append("-");
+        sb.append(UtilAll.getPid()).append("-");
+        sb.append(System.nanoTime());
+        return sb.toString().replace(".", "-");
+    }
+
+    private String createInstance(String servers) {
+        String[] serversArray = servers.split(";");
+        List<String> serversList = new ArrayList<String>();
+        for (String server : serversArray) {
+            if (!serversList.contains(server)) {
+                serversList.add(server);
+            }
+        }
+        Collections.sort(serversList);
+        return String.valueOf(serversList.toString().hashCode());
+    }
+
+    /**
+     * parse msg queue by queue json
+     *
+     * @param messageQueueStr
+     * @return
+     */
+    private MessageQueue parseMessageQueueList(String messageQueueStr) {
+        List<String> messageQueueStrList = Splitter.on(SEMICOLON).omitEmptyStrings().trimResults().splitToList(messageQueueStr);
+        if (org.apache.commons.collections.CollectionUtils.isEmpty(messageQueueStrList) || messageQueueStrList.size() != 3) {
+            return null;
+        }
+        return new MessageQueue(messageQueueStrList.get(0), messageQueueStrList.get(1), Integer.valueOf(messageQueueStrList.get(2)));
     }
 
     /**
@@ -128,8 +235,8 @@ public class RocketMQEventSubscriber extends EventSubscriber {
         String connectSchema = properties.get(RuntimerConfigDefine.CONNECT_SCHEMA);
         schema = StringUtils.isNotEmpty(connectSchema) ? JSON.parseObject(connectSchema, Schema.class) : null;
         byte[] body = messageExt.getBody();
-        RecordPartition recordPartition = listenerFactory.convertToRecordPartition(messageExt.getTopic(), messageExt.getBrokerName(), messageExt.getQueueId());
-        RecordOffset recordOffset = listenerFactory.convertToRecordOffset(messageExt.getQueueOffset());
+        RecordPartition recordPartition = convertToRecordPartition(messageExt.getTopic(), messageExt.getBrokerName(), messageExt.getQueueId());
+        RecordOffset recordOffset = convertToRecordOffset(messageExt.getQueueOffset());
         String bodyStr = new String(body, StandardCharsets.UTF_8);
         sinkRecord = new ConnectRecord(recordPartition, recordOffset, timestamp, schema, bodyStr);
         KeyValue keyValue = new DefaultKeyValue();
@@ -141,6 +248,22 @@ public class RocketMQEventSubscriber extends EventSubscriber {
         }
         sinkRecord.addExtension(keyValue);
         return sinkRecord;
+    }
+
+    private RecordPartition convertToRecordPartition(String topic, String brokerName, int queueId) {
+        Map<String, String> map = new HashMap<>();
+        map.put("topic", topic);
+        map.put("brokerName", brokerName);
+        map.put("queueId", queueId + "");
+        RecordPartition recordPartition = new RecordPartition(map);
+        return recordPartition;
+    }
+
+    private RecordOffset convertToRecordOffset(Long offset) {
+        Map<String, String> offsetMap = new HashMap<>();
+        offsetMap.put(QUEUE_OFFSET, offset + "");
+        RecordOffset recordOffset = new RecordOffset(offsetMap);
+        return recordOffset;
     }
 
     /**
@@ -162,7 +285,5 @@ public class RocketMQEventSubscriber extends EventSubscriber {
     private void unSubscribe(String topic) {
         pullConsumer.unsubscribe(topic);
     }
-
-
 
 }
